(refactor) 결과값을 따로 계산해주는 class 생성 및 테스트
